+++ apps/web/src/components/user-dropdown.tsx
@@ -1,14 +1,14 @@
+import { authClient } from "@/lib/auth-client";
+import { Button } from "@heroui/button";
 import {
   Dropdown,
-  DropdownTrigger,
-  DropdownMenu,
   DropdownItem,
+  DropdownMenu,
+  DropdownTrigger,
 } from "@heroui/dropdown";
-import { Button } from "@heroui/button";
-import { ChevronDownIcon } from "./data-grid";
-import { authClient } from "@/lib/auth-client";
 import { useRouter } from "@tanstack/react-router";
-
+import { LogOut, Settings } from "lucide-react";
+import { ChevronDownIcon } from "./data-grid";
 
 export default function UserDropdown() {
   const { data: session, isPending } = authClient.useSession();
@@ -17,7 +17,7 @@ export default function UserDropdown() {
   const onSettingsClick = () => {
     // Handle settings navigation
     router.navigate({ to: "/settings" });
-  }
+  };
 
   const onLogoutClick = () => {
     // Handle logout logic
@@ -29,22 +29,34 @@ export default function UserDropdown() {
       <DropdownTrigger>
         <Button
           variant="light"
-          endContent={<ChevronDownIcon className="text-small text-default-500" />}
+          endContent={
+            <ChevronDownIcon className="text-small text-default-500" />
+          }
         >
           {isPending ? "Loading..." : session?.user?.name}
         </Button>
       </DropdownTrigger>
       <DropdownMenu aria-label="User actions">
         <DropdownItem key="settings" onPress={onSettingsClick}>
-          Settings
+          <div className="flex items-center gap-2">
+            <div>
+              <Settings size={20} />
+            </div>
+            <div>Settings</div>
+          </div>
         </DropdownItem>
-        <DropdownItem 
-          key="logout" 
-          className="text-danger" 
+        <DropdownItem
+          key="logout"
+          className="text-danger"
           color="danger"
           onPress={onLogoutClick}
         >
-          Logout
+          <div className="flex items-center gap-2">
+            <div>
+              <LogOut size={20} />
+            </div>
+            <div>Logout</div>
+          </div>
         </DropdownItem>
       </DropdownMenu>
     </Dropdown>
