+++ apps/web/src/routes/index.tsx
@@ -1,7 +1,4 @@
 import DataGrid from "@/components/data-grid.tsx";
-import FileUploader from "@/components/file-upload.tsx";
-import { useTRPC } from "@/utils/trpc.ts";
-import { useQuery } from "@tanstack/react-query";
 import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/")({
@@ -9,15 +6,6 @@ export const Route = createFileRoute("/")({
 });
 
 function Index() {
-  const trpc = useTRPC();
-  const { data = [], isLoading } = useQuery(
-    trpc.file.getUploadedFiles.queryOptions(),
-  );
-
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
-
   return (
     <div className="p-4">
       <DataGrid />
